fix: resolve issue with data parsing after updating
