+++ utils/db.ts
@@ -23,10 +23,10 @@ export const authAdapter = PrismaAdapter(prisma);
 
 export async function getUsersPosts(userId: string) {
   const user = await prisma.user.findUnique({
-    cacheStrategy: {
-      swr: 120,
-      ttl: 120,
-    },
+    // cacheStrategy: {
+    //   swr: 120,
+    //   ttl: 120,
+    // },
     where: { id: userId },
     select: {
       id: true,
@@ -61,10 +61,10 @@ export async function getPost(slug: string): Promise<
 > {
   try {
     const data = await prisma.post.findUnique({
-      cacheStrategy: {
-        swr: 240,
-        ttl: 240,
-      },
+      // cacheStrategy: {
+      //   swr: 240,
+      //   ttl: 240,
+      // },
       where: { slug },
       select: {
         id: true,
